feat(cli)!: remove setup/exec/debug-doctor/import-legacy/report-legacy shims
Final batch of the 0.8.0 breaking removal. Drops:
- flaker setup init (use: flaker init)
- flaker exec run / exec affected (use: flaker run)
- flaker debug doctor (use: flaker doctor)
- flaker import report <file> / import parquet <dir> (use: flaker import <file>)
- flaker report summary / diff / aggregate (use: flaker report <file> --<mode>)
- flaker kpi top-level alias (use: flaker status)

execRunAction and RUN_COMMAND_HELP moved from src/cli/categories/exec.ts
to src/cli/commands/run.ts so they survive the category deletion.

setupInitAction moved from src/cli/categories/setup.ts
to src/cli/commands/setup/init.ts alongside runInit.

BREAKING CHANGE: scripts using the shims above now exit non-zero.
Migration unchanged from 0.7.0 — see docs/migration-0.6-to-0.7.md.
