+++ src/cli/commands/setup/init.ts
@@ -114,6 +114,33 @@ jobs:
 `;
 }
 
+import { basename } from "node:path";
+import { detectRepoInfo } from "../../core/git.js";
+
+const VALID_ADAPTERS = ["playwright", "vitest", "jest", "junit"] as const;
+const VALID_RUNNERS = ["vitest", "playwright", "jest", "actrun"] as const;
+
+export async function setupInitAction(opts: { owner?: string; name?: string; adapter?: string; runner?: string }): Promise<void> {
+  if (opts.adapter && !VALID_ADAPTERS.includes(opts.adapter as typeof VALID_ADAPTERS[number])) {
+    console.error(`Error: unknown adapter "${opts.adapter}". Valid: ${VALID_ADAPTERS.join(", ")}`);
+    process.exit(1);
+  }
+  if (opts.runner && !VALID_RUNNERS.includes(opts.runner as typeof VALID_RUNNERS[number])) {
+    console.error(`Error: unknown runner "${opts.runner}". Valid: ${VALID_RUNNERS.join(", ")}`);
+    process.exit(1);
+  }
+  const cwd = process.cwd();
+  const detected = detectRepoInfo(cwd);
+  const owner = opts.owner ?? detected?.owner ?? "local";
+  const name = opts.name ?? detected?.name ?? basename(cwd);
+  runInit(cwd, { owner, name, adapter: opts.adapter, runner: opts.runner });
+  if (!detected && !opts.owner) {
+    console.log(`Initialized flaker.toml (${owner}/${name}) — no git remote found, using defaults`);
+  } else {
+    console.log(`Initialized flaker.toml (${owner}/${name})`);
+  }
+}
+
 export function runInit(
   dir: string,
   opts: { owner: string; name: string; adapter?: string; runner?: string },
