+++ tests/cli/help.test.ts
@@ -17,10 +17,10 @@ describe("CLI help", () => {
     expect(help).toContain("Advanced:");
   });
 
-  it("shows exec run help with --dry-run and --explain flags", () => {
+  it("shows run help with --dry-run and --explain flags", () => {
     const program = createProgram();
-    const execCmd = program.commands.find((command) => command.name() === "exec");
-    const runCmd = execCmd?.commands.find((command) => command.name() === "run");
+    // exec category removed in 0.8.0 — use top-level run command.
+    const runCmd = program.commands.find((command) => command.name() === "run");
     const runHelp = runCmd?.helpInformation();
     // gate/quarantine commands removed in 0.8.0 — lookups deleted.
     const opsCmd = program.commands.find((command) => command.name() === "ops");
@@ -28,8 +28,7 @@ describe("CLI help", () => {
     const opsIncidentHelp = opsCmd?.commands.find((command) => command.name() === "incident")?.helpInformation();
     const opsWeeklyHelp = opsCmd?.commands.find((command) => command.name() === "weekly")?.helpInformation();
     // analyze subcommands (eval, bundle, flaky-tag) removed in 0.8.0 — lookups deleted.
-    const importCmd = program.commands.find((command) => command.name() === "import");
-    const importReportHelp = importCmd?.commands.find((command) => command.name() === "report")?.helpInformation();
+    // import report subcommand removed in 0.8.0 — use top-level import <file>.
 
     expect(runHelp).toContain("--dry-run");
     expect(runHelp).toContain("--explain");
@@ -46,6 +45,6 @@ describe("CLI help", () => {
     expect(opsWeeklyHelp).toContain("--window-days");
     expect(opsWeeklyHelp).toContain("--json");
     // evalHelp, bundleHelp, flakyTagHelp assertions removed — commands dropped in 0.8.0.
-    expect(importReportHelp).toContain("--source");
+    // importReportHelp assertion removed — import report subcommand dropped in 0.8.0.
   });
 });
